move the test constants to the test directory rather than source directory
